+++ includes/widgets/shop-demo/article-browser-widget/spec/spec_runner.js
@@ -9,6 +9,8 @@
       title: 'ArticleBrowserWidget Specification',
       tests: [
          'article-browser-widget.spec'
-      ]
+      ],
+      testRunner: 'laxar-testing',
+      jasmineMajorVersion: 2
    };
 } )( this );